+++ packages/react-native-sdk/expo-config-plugin/src/withMainActivity.ts
@@ -20,6 +20,7 @@ const withStreamVideoReactNativeSDKMainActivity: ConfigPlugin<ConfigProps> = (
           config.modResults.contents,
           [
             'com.streamvideo.reactnative.StreamVideoReactNative',
+            'android.os.Build',
             'android.util.Rational',
             'androidx.lifecycle.Lifecycle',
             'android.app.PictureInPictureParams',
@@ -56,7 +57,7 @@ function addOnPictureInPictureModeChanged(contents: string) {
   @Override
   public void onPictureInPictureModeChanged(boolean isInPictureInPictureMode) {
     super.onPictureInPictureModeChanged(isInPictureInPictureMode);
-    if (getLifecycle().getCurrentState() == Lifecycle.State.CREATED) {
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O && getLifecycle().getCurrentState() == Lifecycle.State.CREATED) {
       // when user clicks on Close button of PIP
       finishAndRemoveTask();
     } else {